Fix rendering of mimir-distributed 4.3.0 Helm chart for OpenShift with Helm 3.17.1
Helm 3.17.1 has fixed a long-standing bug which affects how sub-chart
values are merged in https://github.com/helm/helm/pull/13654. The
Commodore 1.27.3 image uses Helm 3.17.1.

Somehow, the fix breaks the suggested configuration for the
mimir-distributed chart when setting `rbac.type=scc`, since the `null`
value overrides for the sub-chart fields are now explicitly visible as
fields with value `null`.

This breaks the validation logic which checks the presence of the fields
We adjust the validation by copying a fixed `validate.yaml` into the
downloaded helm chart. However, this now injects `null` values in the
rendered manifest, so we also add a postprocessing filter to delete
those nulls.